refactor: move auth data provider interface
