+++ src/adapters/auth/index.ts
@@ -4,6 +4,6 @@
  * This file exports all auth adapter related types and functions.
  */
 
-export * from '@/core/auth/IAuthDataProvider';
+export * from './interfaces';
 export * from './factory';
 export * from './supabase-auth-provider';
