+++ src/core/auth/index.ts
@@ -7,7 +7,6 @@
 
 // Export all interfaces
 export * from './interfaces';
-export * from './IAuthDataProvider';
 
 // Export all models
 export * from './models';
