+++ src/services/auth/default-auth.service.ts
@@ -9,7 +9,7 @@ import {
   AuthService,
   AuthState
 } from '@/core/auth/interfaces';
-import type { IAuthDataProvider } from '@/core/auth/IAuthDataProvider';
+import type { IAuthDataProvider } from '@/adapters/auth/interfaces';
 import { 
   AuthResult, 
   LoginPayload,
